requirements.txt
kafka-python
websocket-client
pykx
numpy
pandas
plotly